migrate from bun to pnpm/tsdown/vitest
I'd understand if you do not want to merge this as it was probably your intentions
to experiment with bun.

However, let me give you a few reasons to consider this:

- bun does not work on windows
- bun integration with IDEs is not great
- a lot of companies (wink) use pnpm as their default package manager, making contributions
easier

In contrast, we can now run tests directly from the IDE thanks to vitest.

Tsdown produces a very similar output to what bun produces and with similar build performance.

Finally, if we compromise on using the built files as dependencies, we can run the examples
directly with node now.